refactor: move `lib.rs` contents to `anonymize.rs`
To prepare for the "Deanonymizer" that will be added in the future.
